+++ src/settings/MappingProfiles/shapes/index.js
@@ -1,1 +1,1 @@
-export * from './mappingProfilesShape';
+export * from './mappingProfileShape';