+++ test/bigtest/interactors/settings/index.js
@@ -1,1 +1,2 @@
 export { default as SettingsSectionsPaneInteractor } from './SettingsSectionsPaneInteractor';
+export * from './ProfileDetailsActionMenuInteractor';