+++ test/bigtest/tests/units/settings/JopProfileDetails/interactors/JobProfileDetailsInteractor.js
@@ -2,12 +2,12 @@ import { interactor } from '@bigtest/interactor';
 
 import { AccordionSetInteractor } from '@folio/stripes-components/lib/Accordion/tests/interactor';
 import MetaSectionInteractor from '@folio/stripes-components/lib/MetaSection/tests/interactor';
-import ButtonInteractor from '@folio/stripes-components/lib/Button/tests/interactor';
 import {
   FullScreenViewInteractor,
   PreloaderInteractor,
 } from '@folio/stripes-data-transfer-components/interactors';
 
+import { ProfileDetailsActionMenuInteractor } from '../../../../../interactors';
 import { SummaryContentInteractor } from './SummaryContentInteractor';
 
 @interactor
@@ -19,7 +19,5 @@ export class JobProfileDetailsInteractor {
   summary = new SummaryContentInteractor();
   preloader = new PreloaderInteractor();
   metadata = new MetaSectionInteractor();
-  editProfileButton = new ButtonInteractor('[data-test-edit-profile-button]');
-  duplicateProfileButton = new ButtonInteractor('[data-test-duplicate-profile-button]');
-  deleteProfileButton = new ButtonInteractor('[data-test-delete-profile-button]');
+  actionMenu = new ProfileDetailsActionMenuInteractor();
 }